-------------------------------------------------------------------------------- Changeset Id: 1875191 Date: 6/18/2010 12:12:42 PM
Fixes a stack overflow on 64-bit w/ deeply nested conditionals.  When we have a large conditional we now generate 

If cond0:
                Body
                Goto end
If cond1:
                Body
                Goto end

else_body
end:

instead of:

if cond0:
                body
else:
                if cond1:
                                body
                else:
                                else_body

That keeps the conditional flat so we don’t stack overflow.  We still generate the old way for small conditionls because it’s simpler to construct.




(Shelveset: FixStackOverflowFinal;REDMOND\dinov | SNAP CheckinId: 10949)