-------------------------------------------------------------------------------- Changeset Id: 1895157 Date: 6/29/2010 10:57:09 AM
Fixes a bug w/ the offsets we’re reporting for tokenization inside of VisualStudio.  We aren’t respecting initial location and we’re not clearing _newLines when re-initializing the tokenizer.  The parser is also failing to set the span on the function associated w/ a lambda expression.



(Shelveset: FixNewParserBug;REDMOND\dinov | SNAP CheckinId: 11021)