-------------------------------------------------------------------------------- Changeset Id: 1967242 Date: 8/10/2010 9:57:20 PM
Adds a new way of implementing built-in modules by subclassing PythonBuiltinModule class.  This enables having a module which has an instance associated with it rather than having a bunch of static method definitions.  We’ll create the instance for the PythonContext when the module gets imported.  The instance can include various fields which enable quick access to the module.

This was requested to help get fast module access for Cython-like support for IronPython.





(Shelveset: NewBuiltinModulefinal;REDMOND\dinov | SNAP CheckinId: 11255)